pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=10.0.0",
    "python-pptx==0.6.23",
    "toml>=0.10.2",
]
